ossl_pkey_ec.c: index types
* ext/openssl/ossl_pkey_ec.c (ossl_ec_point_mul): fix index types.
  Array length is long, not int.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@55052 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/openssl/ossl_pkey_ec.c
@@ -1543,7 +1543,7 @@ static VALUE ossl_ec_point_mul(int argc, VALUE *argv, VALUE self)
 	 * bignums | arg1[0] | arg1[1] | arg1[2] | ...
 	 * points  | self    | arg2[0] | arg2[1] | ...
 	 */
-	int i, num;
+	long i, num;
 	VALUE tmp_p, tmp_b;
 	const EC_POINT **points;
 	const BIGNUM **bignums;
